Add new feature. Import local data such as services, pickup-points and lockers in one single action.
