+++ Controller/Adminhtml/PickupPoint/Refresh.php
@@ -3,60 +3,27 @@
 namespace SamedayCourier\Shipping\Controller\Adminhtml\PickupPoint;
 
 use Magento\Backend\App\Action;
-use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\Framework\Exception\NoSuchEntityException;
 use Sameday\Exceptions\SamedaySDKException;
-use Sameday\Requests\SamedayGetPickupPointsRequest;
-use Sameday\Sameday;
-use SamedayCourier\Shipping\Api\Data\PickupPointInterface;
-use SamedayCourier\Shipping\Api\Data\PickupPointInterfaceFactory;
-use SamedayCourier\Shipping\Api\PickupPointRepositoryInterface;
-use SamedayCourier\Shipping\Helper\ApiHelper;
+use SamedayCourier\Shipping\Helper\LocalDataImporter;
 
 class Refresh extends Action
 {
     /**
-     * @var ApiHelper
+     * @var LocalDataImporter
      */
-    private $apiHelper;
+    private $localDataImporter;
 
     /**
-     * @var ScopeConfigInterface
-     */
-    private $config;
-
-    /**
-     * @var PickupPointInterfaceFactory
-     */
-    private $pickupPointFactory;
-
-    /**
-     * @var PickupPointRepositoryInterface
-     */
-    private $pickupPointRepository;
-
-    /**
-     * Refresh constructor.
-     *
-     * @param ApiHelper $apiHelper
-     * @param ScopeConfigInterface $config
-     * @param PickupPointInterfaceFactory $pickupPointFactory
-     * @param PickupPointRepositoryInterface $pickupPointRepository
      * @param Action\Context $context
+     * @param LocalDataImporter $localDataImporter
      */
     public function __construct(
-        ApiHelper $apiHelper,
-        ScopeConfigInterface $config,
-        PickupPointInterfaceFactory $pickupPointFactory,
-        PickupPointRepositoryInterface $pickupPointRepository,
-        Action\Context $context
+        Action\Context $context,
+        LocalDataImporter $localDataImporter
     ) {
         parent::__construct($context);
 
-        $this->apiHelper = $apiHelper;
-        $this->config = $config;
-        $this->pickupPointFactory = $pickupPointFactory;
-        $this->pickupPointRepository = $pickupPointRepository;
+        $this->localDataImporter = $localDataImporter;
     }
 
     /**
@@ -64,68 +31,9 @@ class Refresh extends Action
      */
     public function execute()
     {
-        $sameday = new Sameday($this->apiHelper->initClient());
-        $isTesting = (bool) $this->config->getValue('carriers/samedaycourier/testing');
-
-        $remotePickupPoints = [];
-        $page = 1;
-        do {
-            $request = new SamedayGetPickupPointsRequest();
-            $request->setPage($page++);
-            try {
-                $pickUpPoints = $sameday->getPickupPoints($request);
-            } catch (\Exception $e) {
-                $this->messageManager->addErrorMessage(sprintf(__('Communication error: %s'), $e->getMessage()));
-
-                return $this->_redirect('samedaycourier_shipping/pickuppoint/index');
-            }
-
-            foreach ($pickUpPoints->getPickupPoints() as $pickupPointObject) {
-                try {
-                    $pickupPoint = $this->pickupPointRepository->getBySamedayId($pickupPointObject->getId(), $isTesting);
-                } catch (NoSuchEntityException $e) {
-                    $pickupPoint = null;
-                }
-
-                if (!$pickupPoint) {
-                    // Pickup point not found, add it.
-                    $pickupPoint = $this->pickupPointFactory->create();
-                }
-
-                $pickupPoint
-                    ->setSamedayId($pickupPointObject->getId())
-                    ->setSamedayAlias($pickupPointObject->getAlias())
-                    ->setIsTesting($isTesting)
-                    ->setCity($pickupPointObject->getCity()->getName())
-                    ->setCounty($pickupPointObject->getCounty()->getName())
-                    ->setAddress($pickupPointObject->getAddress())
-                    ->setContactPersons($pickupPointObject->getContactPersons())
-                    ->setIsDefault($pickupPointObject->isDefault());
-
-                $this->pickupPointRepository->save($pickupPoint);
-
-                // Save as current pickup points.
-                $remotePickupPoints[] = $pickupPointObject->getId();
-            }
-        } while ($page <= $pickUpPoints->getPages());
-
-
-        // Build array of local pickup points.
-        $localPickupPoints = array_map(
-            function (PickupPointInterface $pickupPoint) {
-                return array(
-                    'id' => $pickupPoint->getId(),
-                    'sameday_id' => $pickupPoint->getSamedayId()
-                );
-            },
-            $this->pickupPointRepository->getListByTesting($isTesting)->getItems()
-        );
-
-        // Delete local pickup points that aren't present in remote pickup points anymore.
-        foreach ($localPickupPoints as $localPickupPoint) {
-            if (!in_array($localPickupPoint['sameday_id'], $remotePickupPoints, false)) {
-                $this->pickupPointRepository->deleteById($localPickupPoint['id']);
-            }
+        $import = $this->localDataImporter->importPickupPoints();
+        if (false === $import->isSucceed()) {
+            $this->messageManager->addErrorMessage(sprintf(__('Communication error: %s'), $import->getMessage()));
         }
 
         return $this->_redirect('samedaycourier_shipping/pickuppoint/index');
