+++ view/adminhtml/layout/adminhtml_system_config_edit.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <head>
+        <script src="SamedayCourier_Shipping::js/sameday_import_local_data.js"/>
+    </head>
+</page>
